Move chat create send into chat service
